% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTM_marker_finder.R
\name{.PTM_MarkerFinder_}
\alias{.PTM_MarkerFinder_}
\title{\code{.PTM_MarkerFinder_} a helper function if peakplot option is FALSE .}
\usage{
.PTM_MarkerFinder_(peptideSequence, modification, modi_name)
}
\arguments{
\item{peptideSequence}{peptide sequence}

\item{modification}{contain modification information , intensity of ion, amino acide that is modified}

\item{modi_name}{name of modification}
}
\description{
\code{.PTM_MarkerFinder_} a helper function if peakplot option is FALSE .
}
