% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTM_marker_finder.R
\name{.PTM_MarkerFinder_writeMGF}
\alias{.PTM_MarkerFinder_writeMGF}
\title{\code{.PTM_MarkerFinder_writeMGF} a helper function to find marker ion}
\usage{
.PTM_MarkerFinder_writeMGF(spec, mgfFilename, pattern = -1)
}
\arguments{
\item{spec}{spectrum of peptides}

\item{mgfFilename}{file name of the MGF}

\item{pattern}{indicator for pattern plot}
}
\description{
\code{.PTM_MarkerFinder_writeMGF} a helper function to find marker ion
}
