% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTM_marker_finder.R
\name{.PTM_MarkerFinder_writeMGF_Header}
\alias{.PTM_MarkerFinder_writeMGF_Header}
\title{\code{.PTM_MarkerFinder_writeMGF_Header} a helper function to write the mgf header  .}
\usage{
.PTM_MarkerFinder_writeMGF_Header(mgfFilename)
}
\arguments{
\item{mgfFilename}{file name of the MGF}
}
\description{
\code{.PTM_MarkerFinder_writeMGF_Header} a helper function to write the mgf header  .
}
