% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msptm.R
\name{get_modified}
\alias{get_modified}
\title{\code{get_modified} extract required modification information (modification location,
 modified sequence, type of modification) from protein information}
\usage{
get_modified(uid, modification, result.file)
}
\arguments{
\item{uid}{prot.uid attribute from result.file's peptide slot}

\item{modification}{interested modifications}

\item{result.file}{full information of x!tandem search engine return}
}
\value{
A dataframe that contain protein's uid, sequence, location of modification, 
type of modification.
 
 modification<-data.frame("type"=c("Carbamidomethyl","Oxidation"),
 "monomass"=c(57.022, 16.0), "AA"=c("C","M"))
get_modified(uid=4,modification)
}
\description{
\code{get_modified} extract required modification information (modification location,
 modified sequence, type of modification) from protein information
}
