% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{intensity_plot}
\alias{intensity_plot}
\title{\code{intensity_plot} output graph of intensity of amino acid.}
\usage{
intensity_plot(data, modification, mZmarker_ions, search_engine)
}
\arguments{
\item{data}{mass spetrometry information for the peptide}

\item{modification}{contain modification information , intensity of ion, amino acide that is modified}

\item{mZmarker_ions}{maker ion}

\item{search_engine}{can be Mascot or Tandem}
}
\description{
\code{intensity_plot} output graph of intensity of amino acid.
}
\examples{
modification<-data.frame("type"=c("Carbamidomethyl","Oxidation"),
"monomass"=c(57.022, 16.0), "AA"=c("C","M"))
result.file <- "/Users/yufei/Desktop/2018fall/BCB410/MSPTM/inst/extdata/output_mouse.2018_12_04_19_57_17.t.xml"
uids<-c(12,2,731)
library(rTANDEM)
result <- GetResultsFromXML(result.file)
data<-tandem_get_data(result,modification,uids)
intensity_plot(data,modification,mZmarker_ions, search_engine="Tandem")
}
