% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data.R
\name{mascot_data}
\alias{mascot_data}
\title{refereence: perl script comes from package protViz get rdatafile from a mascot output file}
\usage{
mascot_data(iput_file, oput_file)
}
\arguments{
\item{iput_file}{input file that is received from mascot}

\item{oput_file}{a file that is readable by the protviz package}
}
\description{
refereence: perl script comes from package protViz get rdatafile from a mascot output file
}
