% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msptm.R
\name{tandem_get_data}
\alias{tandem_get_data}
\title{\code{tandem_get_data} extract required data(modification, peptide sequence, mass 
to charge) from rtandem output file.}
\usage{
tandem_get_data(result, modification, uids)
}
\arguments{
\item{result}{a output file from rtandem package that contain mass to charge,
peptide sequence, protein information, modification}

\item{modification}{modification information that are investigated}

\item{uids}{proteins id that you want to investigate}
}
\value{
A numeric vector.
}
\description{
transform the rtandem output file to a data variables that can be accepted by 
PTM_finderfunction.
}
\examples{
modification<-data.frame("type"=c("Carbamidomethyl","Oxidation")
,"monomass"=c(57.022, 16.0), "AA"=c("C","M"))
tandem_get_data(result.file,modification)
}
